block: anxiety-iosched: don't use size_t type for starvation
size_t should not be used in this case.

+++ block/anxiety-iosched.c
@@ -14,7 +14,7 @@
 
 struct anxiety_data {
 	struct list_head queue[2];
-	size_t writes_starved;
+	unsigned int writes_starved;
 };
 
 static void anxiety_merged_requests(struct request_queue *q, struct request *rq, struct request *next)
